pyproject.toml
[project]
name = "wilmut-invader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.2",
    "mistletoe>=1.4.0",
    "pygame>=2.6.1",
    "pygbag>=0.9.2",
    "typer>=0.15.1",
]

[project.scripts]
wilmut-invader = "wilmut_invader:main.start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.format]
quote-style = "single"
